Move render function to seperate helper file
