+++ .gitignore
@@ -3,4 +3,4 @@ src/config.json
 report.json
 report.xml
 .env
-test.html
+report.html
